requirements.txt
fastapi 
uvicorn[standard] 
sqlalchemy 
psycopg2-binary 
alembic

python-jose[cryptography] 
passlib[bcrypt] 
python-multipart

pydantic
redis
asyncio
fastapi_limiter
httpx
requests
decouple==0.0.7